Refactor, more decoupled, scope rectified and better single responsibility.
