+++ src/commands/azureServiceOperators/helpers/azureservicehelper.ts
@@ -2,12 +2,79 @@ import * as vscode from 'vscode';
 import * as k8s from 'vscode-kubernetes-tools-api';
 import * as path from 'path';
 import * as fs from 'fs';
-import { getExtensionPath } from '../../utils/host';
+import { getExtensionPath, longRunning } from '../../utils/host';
 import { OperatorSettings } from '../models/operatorSettings';
 import * as tmpfile from '../../utils/tempfile';
+import AksClusterTreeItem from '../../../tree/aksClusterTreeItem';
+import { InstallationResponse } from '../models/installationResponse';
+import * as clusters from '../../utils/clusters';
+import { createASOWebView } from './azureservicehtmlhelper';
 const tmp = require('tmp');
 
-export async function getOperatorsPod(
+export async function startInstallation(
+    panel: vscode.WebviewPanel,
+    extensionPath: string,
+    kubectl: k8s.KubectlV1,
+    installationResponse: InstallationResponse,
+    aksCluster: AksClusterTreeItem,
+    operatorSettingsInfo: OperatorSettings
+) {
+
+    // getKubeconfigYaml handles reporting failure to the user, hence we dont need it here.
+    const clusterKubeConfig = await clusters.getKubeconfigYaml(aksCluster);
+    if (!clusterKubeConfig) return undefined;
+
+    // 1) Azure Service Operator requires self-signed certificates for CRD Conversion Webhooks.
+    installationResponse.installCertManagerResult = await longRunning(`Installing Cert Manager for Azure Service Operator...`,
+        () => installCertManager(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.installCertManagerResult, installationResponse))) return undefined;
+
+    // 2) The cert-manager pods should be running before proceeding to the next step.
+    installationResponse.checkCertManagerRolloutStatusResult = await longRunning(`Checking Cert Manager Rollout Status...`,
+        () => checkCertManagerRolloutStatus(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.checkCertManagerRolloutStatusResult, installationResponse))) return undefined;
+
+    // 3) Install OLM is the pre-requisite of this work, using the apply YAML instructions here: https://github.com/operator-framework/operator-lifecycle-manager/releases/.
+    // Also, page to refer: https://operatorhub.io/operator/azure-service-operator (Click Install button as top of the page)
+    installationResponse.installOlmCrdResult = await longRunning(`Installing Operator Lifecycle Manager CRD resource...`,
+        () => installOlmCrd(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.installOlmCrdResult, installationResponse))) return undefined;
+
+    installationResponse.installOlmResult = await longRunning(`Installing Operator Lifecycle Manager resource...`,
+        () => installOlm(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.installOlmResult, installationResponse))) return undefined;
+
+    installationResponse.installOperatorResult = await longRunning(`Installing Opreator Namespace...`,
+        () => installOperator(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.installOperatorResult, installationResponse))) return undefined;
+
+    // 4) IssuerCert apply with Operator namespace created above.
+    installationResponse.installIssuerCertResult = await longRunning(`Installing the Issuer and Certificate cert-manager resources....`,
+        () => installIssuerCert(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.installIssuerCertResult, installationResponse))) return undefined;
+
+    // 5) Run kubectl apply for azureoperatorsettings.yaml
+    installationResponse.installOperatorSettingsResult = await longRunning(`Installing Azure Service Operator Settings...`,
+        () => installOperatorSettings(kubectl, operatorSettingsInfo, clusterKubeConfig)
+    );
+    if (!isInstallationSuccessfull(panel, extensionPath, installationResponse.installOperatorSettingsResult, installationResponse)) return undefined;
+
+    // 6) Final step: Get the azure service operator pod. - kubectl get pods -n operators
+    installationResponse.getOperatorsPodResult = await longRunning(`Getting Azure Service Operator Pod...`,
+        () => getOperatorsPod(kubectl, clusterKubeConfig)
+    );
+    if (!(isInstallationSuccessfull(panel, extensionPath, installationResponse.getOperatorsPodResult, installationResponse))) return undefined;
+
+    createASOWebView(panel, extensionPath, installationResponse);
+}
+
+async function getOperatorsPod(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -19,7 +86,7 @@ export async function getOperatorsPod(
     return result;
 }
 
-export async function installCertManager(
+async function installCertManager(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -31,7 +98,7 @@ export async function installCertManager(
     return result;
 }
 
-export async function checkCertManagerRolloutStatus(
+async function checkCertManagerRolloutStatus(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -42,7 +109,7 @@ export async function checkCertManagerRolloutStatus(
     return result;
 }
 
-export async function installOlmCrd(
+async function installOlmCrd(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -59,7 +126,7 @@ export async function installOlmCrd(
     }
 }
 
-export async function installOlm(
+async function installOlm(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -71,7 +138,7 @@ export async function installOlm(
     return result;
 }
 
-export async function installOperator(
+async function installOperator(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -83,7 +150,7 @@ export async function installOperator(
     return result;
 }
 
-export async function installIssuerCert(
+async function installIssuerCert(
     kubectl: k8s.KubectlV1,
     clusterKubeConfig: string
 ): Promise<k8s.KubectlV1.ShellResult | undefined> {
@@ -100,7 +167,7 @@ export async function installIssuerCert(
     return result;
 }
 
-export async function installOperatorSettings(
+async function installOperatorSettings(
     kubectl: k8s.KubectlV1,
     operatorSettingInfo: OperatorSettings,
     clusterKubeConfig: string
@@ -122,6 +189,24 @@ export async function installOperatorSettings(
     const failureDescription = "Install operator settings had following error";
     const result = await invokeKubectlCommand(kubectl, clusterKubeConfig, command, failureDescription);
     return result;
+}
+
+function isInstallationSuccessfull(
+    panel: vscode.WebviewPanel,
+    extensionPath: string,
+    installationShellResult: k8s.KubectlV1.ShellResult | undefined,
+    installationResponse: InstallationResponse
+): boolean {
+    let success = true;
+
+    if (!installationShellResult) return false;
+
+    if (installationShellResult.code !== 0) {
+        createASOWebView(panel, extensionPath, installationResponse);
+        success = false;
+    }
+
+    return success;
 }
 
 async function invokeKubectlCommand(
